Make shell script to automate poetry building and distribution of package
Why: It is super easy to forget the commands used to build and distribute
a package since we don't use them every day.

How: Wrote shell commands into a script that can be run by the user
post configuring poetry config paramters properly. The name of the
index to publish can be changed inside the script to publish to a
different index.

Tags: