Add notebook for testing out installation
Why: It seems like a good way to ensure that everything is working
as expected on colab too.

How: The notebook contains some code on testing the installation
of our library on test and normal PyPi. Run the notebook on colab to
ensure that everything works on that platform.

Tags: