Fix classes list not being given to dataset split exports
Why: If common class list isn't given while exporting dataset splits, the
validation/test classes might not match train classes which results in an
error.

How: Using the default_classes stored in the dataset. This list is common
for all the different views of the dataset: train, test, val.

Tags: