Fix inputting of label files to preprocessor
Why: COCO files were being put in a list and were being fed to the preprocessor
which went against how label checking was being performed. Additionally,
this would have been hectic for the user to type every file name
in a list.

How: Replaced the individual file name-based list with the folder name
and processed the folder internally

Tags: