Add multiple changes for inference and Refactor decorator name
Why: Model has to be used for inference. We can use the same preprocessor
class to preprocess the inference dataset. Changes are being made for that.
Additionally, an original decorator wasn't really doing what its name
suggested. Thus, we wanted to ensure that the name aligns with functionality.

How:
* Under Preprocesor, label_folder_path has been rendered optional and
a new self.mode attribute has been added to indicate train or inference
* Decorator name got changed everywhere

Tags: